move login and user function to dedicated file
