+++ Dockerfile
@@ -24,7 +24,8 @@ RUN pip install --no-cache-dir --upgrade pip \
     && pip install --no-cache-dir -r requirements.txt
 
 # Copy application code
-COPY app.py .
+#COPY app.py .
+COPY app /app/
 
 # Create non-root user for security
 RUN adduser --disabled-password --gecos '' appuser \
